+++ handlers/channel_handler.go
@@ -9,7 +9,7 @@ import (
 	"github.com/disgoorg/disgo/gateway"
 )
 
-func gatewayHandlerChannelCreate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventChannelCreate) {
+func gatewayHandlerChannelCreate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventChannelCreate) {
 	client.Caches().AddChannel(event.GuildChannel)
 
 	client.EventManager().DispatchEvent(&events.GuildChannelCreate{
@@ -22,7 +22,7 @@ func gatewayHandlerChannelCreate(client bot.Client, sequenceNumber int, shardID 
 	})
 }
 
-func gatewayHandlerChannelUpdate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventChannelUpdate) {
+func gatewayHandlerChannelUpdate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventChannelUpdate) {
 	oldGuildChannel, _ := client.Caches().Channel(event.ID())
 	client.Caches().AddChannel(event.GuildChannel)
 
@@ -57,7 +57,7 @@ func gatewayHandlerChannelUpdate(client bot.Client, sequenceNumber int, shardID 
 	}
 }
 
-func gatewayHandlerChannelDelete(client bot.Client, sequenceNumber int, shardID int, event gateway.EventChannelDelete) {
+func gatewayHandlerChannelDelete(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventChannelDelete) {
 	client.Caches().RemoveChannel(event.ID())
 
 	client.EventManager().DispatchEvent(&events.GuildChannelDelete{
@@ -70,7 +70,7 @@ func gatewayHandlerChannelDelete(client bot.Client, sequenceNumber int, shardID 
 	})
 }
 
-func gatewayHandlerChannelPinsUpdate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventChannelPinsUpdate) {
+func gatewayHandlerChannelPinsUpdate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventChannelPinsUpdate) {
 	if event.GuildID == nil {
 		client.EventManager().DispatchEvent(&events.DMChannelPinsUpdate{
 			GenericEvent:        events.NewGenericEvent(client, sequenceNumber, shardID),
