+++ handlers/guild_member_handlers.go
@@ -6,7 +6,7 @@ import (
 	"github.com/disgoorg/disgo/gateway"
 )
 
-func gatewayHandlerGuildMemberAdd(client bot.Client, sequenceNumber int, shardID int, event gateway.EventGuildMemberAdd) {
+func gatewayHandlerGuildMemberAdd(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventGuildMemberAdd) {
 	if guild, ok := client.Caches().Guild(event.GuildID); ok {
 		guild.MemberCount++
 		client.Caches().AddGuild(guild)
@@ -23,7 +23,7 @@ func gatewayHandlerGuildMemberAdd(client bot.Client, sequenceNumber int, shardID
 	})
 }
 
-func gatewayHandlerGuildMemberUpdate(client bot.Client, sequenceNumber int, shardID int, event gateway.EventGuildMemberUpdate) {
+func gatewayHandlerGuildMemberUpdate(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventGuildMemberUpdate) {
 	oldMember, _ := client.Caches().Member(event.GuildID, event.User.ID)
 	client.Caches().AddMember(event.Member)
 
@@ -37,7 +37,7 @@ func gatewayHandlerGuildMemberUpdate(client bot.Client, sequenceNumber int, shar
 	})
 }
 
-func gatewayHandlerGuildMemberRemove(client bot.Client, sequenceNumber int, shardID int, event gateway.EventGuildMemberRemove) {
+func gatewayHandlerGuildMemberRemove(client bot.Client, sequenceNumber uint64, shardID int, event gateway.EventGuildMemberRemove) {
 	if guild, ok := client.Caches().Guild(event.GuildID); ok {
 		guild.MemberCount--
 		client.Caches().AddGuild(guild)
@@ -53,7 +53,7 @@ func gatewayHandlerGuildMemberRemove(client bot.Client, sequenceNumber int, shar
 	})
 }
 
-func gatewayHandlerGuildMembersChunk(client bot.Client, _ int, _ int, event gateway.EventGuildMembersChunk) {
+func gatewayHandlerGuildMembersChunk(client bot.Client, _ uint64, _ int, event gateway.EventGuildMembersChunk) {
 	for i := range event.Members {
 		event.Members[i].GuildID = event.GuildID
 	}
